Started work on new rendering system.

+++ LambdaEngine/Components/Rendering/PrimitiveComponents/RectPrimitiveComponent.cs
@@ -6,10 +6,10 @@ namespace LambdaEngine.Components.Rendering.PrimitiveComponents;
 
 [EcsComponent]
 public struct RectPrimitiveComponent(float x, float y) : IEcsComponent {
-    public int ZIndex = 0;
+    public sbyte ZIndex = 0;
     public Vector2 Size = new(x ,y);
 
-    public RectPrimitiveComponent(float x, float y, int zIndex) : this(x, y) {
+    public RectPrimitiveComponent(float x, float y, sbyte zIndex) : this(x, y) {
         ZIndex = zIndex;
     }
 }
