+++ LambdaEngine/Components/Rendering/SpriteComponent.cs
@@ -7,14 +7,14 @@ namespace LambdaEngine.Components.Rendering;
 [EcsComponent]
 public struct SpriteComponent : IEcsComponent {
     internal IntPtr TextureHandle;
-    public int ZIndex;
+    public sbyte ZIndex;
 
     public Texture Texture {
         get => new(TextureHandle);
         set => TextureHandle = value.Handle;
     }
     
-    public SpriteComponent(Texture texture, int zIndex = 0) {
+    public SpriteComponent(Texture texture, sbyte zIndex = 0) {
         TextureHandle = texture.Handle;
         ZIndex = zIndex;
     }
